num (int?) to double

+++ lib/screens/transaction_screens/add_transaction_screen.dart
@@ -39,7 +39,7 @@ class _AddTransactionScreenState extends State<AddTransactionScreen> {
   void getBalance() async {
     DataSnapshot snapshot = await dbBalance.child("user0").get();
     Map balance = snapshot.value as Map;
-    _amount = balance['amount'];
+    _amount = balance['amount'].toDouble();
   }
 
   @override
